Introduce request context and refactor pipeline code
