+++ packages/apollo-cache-control/src/index.ts
@@ -8,7 +8,8 @@ import {
   responsePathAsArray,
 } from 'graphql';
 
-import { GraphQLExtension } from 'graphql-extensions';
+import { GraphQLExtension, GraphQLResponse } from 'graphql-extensions';
+import { Headers } from 'apollo-server-env';
 
 export interface CacheControlFormat {
   version: 1;
@@ -134,6 +135,24 @@ export class CacheControlExtension<TContext = any>
         })),
       },
     ];
+  }
+
+  public willSendResponse?(o: { graphqlResponse: GraphQLResponse }) {
+    if (this.options.calculateHttpHeaders) {
+      const overallCachePolicy = this.computeOverallCachePolicy();
+
+      if (overallCachePolicy) {
+        if (!o.graphqlResponse.http) {
+          o.graphqlResponse.http = { headers: new Headers() };
+        }
+        o.graphqlResponse.http.headers.set(
+          'Cache-Control',
+          `max-age=${
+            overallCachePolicy.maxAge
+          }, ${overallCachePolicy.scope.toLowerCase()}`,
+        );
+      }
+    }
   }
 
   computeOverallCachePolicy(): Required<CacheHint> | undefined {
