+++ packages/apollo-server-plugin-base/package.json
@@ -9,5 +9,12 @@
   "license": "MIT",
   "engines": {
     "node": ">=6"
+  },
+  "devDependencies": {
+    "apollo-server-core": "file:../apollo-server-core",
+    "apollo-server-env": "file:../apollo-server-env"
+  },
+  "peerDependencies": {
+    "graphql": "^0.12.0 || ^0.13.0 || ^14.0.0"
   }
 }
