+++ packages/apollo-server-plugin-base/src/index.ts
@@ -1,10 +1,21 @@
-import { GraphQLRequest } from 'apollo-server-core/src/requestPipelineAPI';
+import {
+  GraphQLRequestContext,
+  GraphQLRequest,
+} from 'apollo-server-core/dist/requestPipelineAPI';
+import { OperationDefinitionNode } from 'graphql';
 
 export abstract class ApolloServerPlugin {
   serverWillStart?(): void;
-  requestDidStart?(): GraphQLRequestListener;
+  requestDidStart?<TContext>(
+    requestContext: GraphQLRequestContext<TContext>,
+  ): GraphQLRequestListener;
 }
 
 export interface GraphQLRequestListener {
   prepareRequest?(request: GraphQLRequest): void;
+  executionDidStart({
+    operation,
+  }: {
+    operation: OperationDefinitionNode;
+  }): void;
 }
