+++ packages/graphql-extensions/package.json
@@ -13,7 +13,8 @@
   "engines": {
     "node": ">=6.0"
   },
-  "dependencies": {
+  "devDependencies": {
+    "apollo-server-core": "file:../apollo-server-core",
     "apollo-server-env": "file:../apollo-server-env"
   },
   "peerDependencies": {
